requirements.txt
python-dotenv~=1.0.1
telebot~=0.0.5
pipenv~=11.9.0
requests~=2.32.0
schedule~=1.1.0
pyTelegramBotAPI==4.12.0